Displays the place and adjust zoom in Google Map.

+++ HomeRental/Models/Rental.cs
@@ -40,8 +40,8 @@ namespace HomeRental.Models
         public string City { get; set; }
         [Required] 
         public string Country { get; set; }
-        public int? Longitude { get; set; }
-        public int? Latitude { get; set; }
+        public double? Longitude { get; set; }
+        public double? Latitude { get; set; }
         public virtual ICollection<Reservation> Reservations { get; set; }
     }
 }
